requirements-macos.txt
# Minimal macOS / Apple Silicon dependencies for PosterAgent.
# Target runtime: Python 3.11
#
# This file keeps the OpenAI-based PosterAgent main path and CLI usable on
# macOS without pulling in Linux/CUDA-only backends.
#
# Intentionally excluded on macOS for the minimal path:
# triton
# xformers==0.0.28.post3
# vllm==0.6.6.post1
# sglang==0.4.2
# compressed-tensors
#
# Optional parser fallback removed from the minimal profile:
# marker-pdf
# FlagEmbedding

python-dotenv==1.0.1
Jinja2==3.1.5
PyYAML==6.0.2

openai==1.59.8
tiktoken==0.7.0
pydantic==2.9.2
pydantic-settings==2.7.1

numpy==1.26.4
scikit-learn==1.6.1
matplotlib==3.10.0

Pillow==10.4.0
pdf2image==1.17.0
lxml==5.3.0
json_repair==0.35.0
Levenshtein==0.26.1
rich==13.9.4
tenacity==9.0.0

requests==2.32.3
duckduckgo_search==6.4.2

python-pptx @ git+https://github.com/Force1ess/python-pptx@dc356685d4d210a10abe1ffab3c21315cdfae63d

docling_core>=2.0.0
docling_parse
docling_ibm_models
